requirements.txt
fastapi
beautifulsoup4
lxml
requests
pillow